refactor: move components to styledComponents
