+++ src/utils/styledComponents.ts
@@ -1,4 +1,5 @@
 import styled from 'styled-components';
+import { NavLink } from 'react-router-dom';
 import svgError from '../public/error.svg';
 import svgSuccess from '../public/success.svg';
 
@@ -53,5 +54,134 @@ export const AlertSuccess = styled.div`
     background-size: 24px 24px;
     background-repeat: no-repeat;
     background-position: center center;
+  }
+`;
+
+export const Container = styled.div`
+  width: 100%;
+  padding-right: 1.5rem;
+  padding-left: 1.5rem;
+  margin-right: auto;
+  margin-left: auto;
+
+  @media (max-width: 992px) {
+    max-width: 93rem;
+  }
+
+  @media (max-width: 768px) {
+    max-width: 72rem;
+  }
+
+  @media (max-width: 576px) {
+    max-width: 54rem;
+  }
+`;
+
+export const Wrapper = styled.div`
+  margin-bottom: 2rem;
+  padding: 1.5rem 2rem 4rem;
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  background-color: rgba(236, 240, 241, .73);
+  border-radius: .5rem;
+`;
+
+export const BackButton = styled(NavLink)`
+  padding: 16px 16px 16px 0;
+  display: inline-flex;
+  flex-wrap: nowrap;
+  align-items: center;
+  font-size: 15px;
+  line-height: 24px;
+  color: #fff;
+  text-decoration: none;
+  transition-property: background-color, color;
+  transition-duration: 0.2s;
+  transition-timing-function: ease-in-out;
+  cursor: pointer;
+
+  & span {
+    margin-left: 5px;
+    color: #573ba3;
+  }
+
+  & svg {
+    fill: #000;
+    vertical-align: bottom;
+    transform-origin: center;
+    transform: rotate(180deg);
+    border-radius: 50%;
+    background-color: #ccc;
+  }
+
+  &:hover {
+    color: rgba(255,255,255,0.8);
+  }
+
+  @media only screen and (max-width: 767px) and (min-width: 0) {
+    padding: 20px 16px;
+    color: rgba(0,0,0,0.48);
+
+    &:hover {
+      color: rgba(0,0,0,0.8);
+    }
+
+    & svg {
+      background-color: rgba(0,0,0,0.1);
+    }
+  }
+`;
+
+export const Row = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+`;
+
+export const AddBtn = styled.button`
+  padding: 0 1rem;
+  margin-right: 0.4rem;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  align-self: flex-end;
+  font-size: 1.2rem;
+  line-height: 1.5;
+  font-weight: 400;
+  text-align: center;
+  vertical-align: middle;
+  color: #fff;
+  border: 1px solid #95a5a6;
+  background-color: #95a5a6;
+  border-radius: .4rem;
+  user-select: none;
+  cursor: pointer;
+  transition: color .15s ease-in-out, background-color .15s ease-in-out, border-color .15s ease-in-out, box-shadow .15s ease-in-out;
+
+  &:hover:not(:disabled) {
+    background-color: #809395;
+    border-color: #798d8f;
+  }
+
+  &:disabled {
+    cursor: auto;
+  }
+
+  & svg {
+    margin: 0.8rem;
+    width: 2rem;
+    height: 2rem;
+    color: #212529;
+    font-size: 2.4rem;
+    font-weight: 400;
+    line-height: 1.5;
+    cursor: pointer;
+    user-select: none;
+  }
+
+  @media (max-width: 415px) {
+    padding: 0 0.2rem;
+    margin-right: 0.2rem;
   }
 `;
